feat: teach common.sh about building LLVM IR
Since we now produce unoptimized LLVM IR, we need to invoke
the LLVM pass pipeline and also build the object file.

Currently, this fails for me since I don't have a libstdc++.
Need to install it.

The global question remains, though: how do we get our hands
on an appropriate `clang`? It appears that `leanc` executes `cc`,
but there is no guarantee that `cc` points to the correct bundled
`clang`.

Currently testing with:
$ ./test_single.sh partial.lean